improved solution of 1647th problem
I used a greedy algorithm.
First, I obtained the frequencies of characters by observing each character in the string and storing them in an array with their respective frequencies.
Next, for each value in the array, I attempted to find a free place without collisions by using a set that already contained existing frequency values.
If a collision was detected, I removed one of the current characters (decreasing the frequency while increasing the result value).